Refactor executeInChildContext related logic to separate class
