+++ src/main/java/com/hubspot/jinjava/lib/tag/eager/EagerIfTag.java
@@ -11,9 +11,9 @@ import com.hubspot.jinjava.lib.tag.IfTag;
 import com.hubspot.jinjava.tree.Node;
 import com.hubspot.jinjava.tree.TagNode;
 import com.hubspot.jinjava.tree.parse.NoteToken;
+import com.hubspot.jinjava.util.EagerContextWatcher;
 import com.hubspot.jinjava.util.EagerExpressionResolver.EagerExpressionResult;
 import com.hubspot.jinjava.util.EagerReconstructionUtils;
-import com.hubspot.jinjava.util.EagerReconstructionUtils.EagerChildContextConfig;
 import com.hubspot.jinjava.util.LengthLimitingStringBuilder;
 import java.util.HashSet;
 import java.util.Set;
@@ -54,15 +54,15 @@ public class EagerIfTag extends EagerTagDecorator<IfTag> {
     );
 
     result.append(
-      EagerReconstructionUtils
+      EagerContextWatcher
         .executeInChildContext(
           eagerInterpreter ->
             EagerExpressionResult.fromString(
               eagerRenderBranches(tagNode, eagerInterpreter, e)
             ),
           interpreter,
-          EagerChildContextConfig
-            .newBuilder()
+          EagerContextWatcher
+            .EagerChildContextConfig.newBuilder()
             .withForceDeferredExecutionMode(true)
             .build()
         )
@@ -108,14 +108,14 @@ public class EagerIfTag extends EagerTagDecorator<IfTag> {
       int branchEnd = findNextElseToken(tagNode, branchStart);
       if (!definitelyDrop) {
         int finalBranchStart = branchStart;
-        EagerExecutionResult result = EagerReconstructionUtils.executeInChildContext(
+        EagerExecutionResult result = EagerContextWatcher.executeInChildContext(
           eagerInterpreter ->
             EagerExpressionResult.fromString(
               evaluateBranch(tagNode, finalBranchStart, branchEnd, interpreter)
             ),
           interpreter,
-          EagerChildContextConfig
-            .newBuilder()
+          EagerContextWatcher
+            .EagerChildContextConfig.newBuilder()
             .withForceDeferredExecutionMode(true)
             .build()
         );
