+++ src/main/java/com/hubspot/jinjava/lib/tag/eager/EagerStateChangingTag.java
@@ -6,9 +6,9 @@ import com.hubspot.jinjava.lib.tag.FlexibleTag;
 import com.hubspot.jinjava.lib.tag.Tag;
 import com.hubspot.jinjava.tree.TagNode;
 import com.hubspot.jinjava.tree.parse.TagToken;
+import com.hubspot.jinjava.util.EagerContextWatcher;
 import com.hubspot.jinjava.util.EagerExpressionResolver.EagerExpressionResult;
 import com.hubspot.jinjava.util.EagerReconstructionUtils;
-import com.hubspot.jinjava.util.EagerReconstructionUtils.EagerChildContextConfig;
 import org.apache.commons.lang3.StringUtils;
 
 public class EagerStateChangingTag<T extends Tag> extends EagerTagDecorator<T> {
@@ -37,13 +37,13 @@ public class EagerStateChangingTag<T extends Tag> extends EagerTagDecorator<T> {
 
     if (!tagNode.getChildren().isEmpty()) {
       result.append(
-        EagerReconstructionUtils
+        EagerContextWatcher
           .executeInChildContext(
             eagerInterpreter ->
               EagerExpressionResult.fromString(renderChildren(tagNode, eagerInterpreter)),
             interpreter,
-            EagerChildContextConfig
-              .newBuilder()
+            EagerContextWatcher
+              .EagerChildContextConfig.newBuilder()
               .withForceDeferredExecutionMode(true)
               .build()
           )
