+++ src/main/java/com/hubspot/jinjava/util/EagerReconstructionUtils.java
@@ -7,12 +7,8 @@ import com.hubspot.jinjava.interpret.Context.Library;
 import com.hubspot.jinjava.interpret.DeferredLazyReference;
 import com.hubspot.jinjava.interpret.DeferredLazyReferenceSource;
 import com.hubspot.jinjava.interpret.DeferredValue;
-import com.hubspot.jinjava.interpret.DeferredValueException;
 import com.hubspot.jinjava.interpret.DisabledException;
 import com.hubspot.jinjava.interpret.JinjavaInterpreter;
-import com.hubspot.jinjava.interpret.JinjavaInterpreter.InterpreterScopeClosable;
-import com.hubspot.jinjava.interpret.LazyExpression;
-import com.hubspot.jinjava.interpret.RevertibleObject;
 import com.hubspot.jinjava.lib.fn.MacroFunction;
 import com.hubspot.jinjava.lib.fn.eager.EagerMacroFunction;
 import com.hubspot.jinjava.lib.tag.AutoEscapeTag;
@@ -23,12 +19,11 @@ import com.hubspot.jinjava.lib.tag.SetTag;
 import com.hubspot.jinjava.lib.tag.eager.DeferredToken;
 import com.hubspot.jinjava.lib.tag.eager.EagerExecutionResult;
 import com.hubspot.jinjava.mode.EagerExecutionMode;
-import com.hubspot.jinjava.objects.collections.PyList;
-import com.hubspot.jinjava.objects.collections.PyMap;
 import com.hubspot.jinjava.objects.serialization.PyishBlockSetSerializable;
 import com.hubspot.jinjava.objects.serialization.PyishObjectMapper;
 import com.hubspot.jinjava.tree.TagNode;
 import com.hubspot.jinjava.tree.parse.TagToken;
+import com.hubspot.jinjava.util.EagerContextWatcher.EagerChildContextConfig;
 import com.hubspot.jinjava.util.EagerExpressionResolver.EagerExpressionResult;
 import java.util.AbstractMap;
 import java.util.Collections;
@@ -44,9 +39,9 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 public class EagerReconstructionUtils {
-  public static final String CANNOT_RECONSTRUCT_MESSAGE = "Cannot reconstruct value";
 
   /**
+   * @deprecated Use {@link EagerContextWatcher#executeInChildContext(Function, JinjavaInterpreter, EagerChildContextConfig)}
    * Execute the specified functions within a protected context.
    * Additionally, if the execution causes existing values on the context to become
    *   deferred, then their previous values will wrapped in a <code>set</code>
@@ -69,6 +64,7 @@ public class EagerReconstructionUtils {
    *  <code>prefixToPreserveState</code> is either blank or a <code>set</code> tag
    *    that preserves the state within the output for a second rendering pass.
    */
+  @Deprecated
   public static EagerExecutionResult executeInChildContext(
     Function<JinjavaInterpreter, EagerExpressionResult> function,
     JinjavaInterpreter interpreter,
@@ -93,334 +89,11 @@ public class EagerReconstructionUtils {
     JinjavaInterpreter interpreter,
     EagerChildContextConfig eagerChildContextConfig
   ) {
-    final Set<String> metaContextVariables = interpreter
-      .getContext()
-      .getMetaContextVariables();
-    final EagerExecutionResult initialResult;
-    final Map<String, Object> speculativeBindings;
-    if (eagerChildContextConfig.checkForContextChanges) {
-      final Set<Entry<String, Object>> entrySet = interpreter.getContext().entrySet();
-      final Map<String, Object> initiallyResolvedHashes = getInitiallyResolvedHashes(
-        entrySet,
-        metaContextVariables
-      );
-      final Map<String, String> initiallyResolvedAsStrings = getInitiallyResolvedAsStrings(
-        interpreter,
-        entrySet,
-        initiallyResolvedHashes
-      );
-      initialResult = applyFunction(function, interpreter, eagerChildContextConfig);
-      speculativeBindings =
-        getAllSpeculativeBindings(
-          interpreter,
-          eagerChildContextConfig,
-          metaContextVariables,
-          initiallyResolvedHashes,
-          initiallyResolvedAsStrings,
-          initialResult.getSpeculativeBindings()
-        );
-    } else {
-      initialResult = applyFunction(function, interpreter, eagerChildContextConfig);
-      speculativeBindings =
-        getBasicSpeculativeBindings(
-          interpreter,
-          eagerChildContextConfig,
-          metaContextVariables,
-          initialResult.getSpeculativeBindings()
-        );
-    }
-    return new EagerExecutionResult(initialResult.getResult(), speculativeBindings);
-  }
-
-  private static EagerExecutionResult applyFunction(
-    Function<JinjavaInterpreter, EagerExpressionResult> function,
-    JinjavaInterpreter interpreter,
-    EagerChildContextConfig eagerChildContextConfig
-  ) {
-    // Don't create new call stacks to prevent hitting max recursion with this silent new scope
-    try (InterpreterScopeClosable c = interpreter.enterNonStackingScope()) {
-      if (eagerChildContextConfig.forceDeferredExecutionMode) {
-        interpreter.getContext().setDeferredExecutionMode(true);
-      }
-      interpreter
-        .getContext()
-        .setPartialMacroEvaluation(eagerChildContextConfig.partialMacroEvaluation);
-      return new EagerExecutionResult(
-        function.apply(interpreter),
-        eagerChildContextConfig.discardSessionBindings
-          ? new HashMap<>()
-          : interpreter.getContext().getSessionBindings()
-      );
-    }
-  }
-
-  private static Map<String, String> getInitiallyResolvedAsStrings(
-    JinjavaInterpreter interpreter,
-    Set<Entry<String, Object>> entrySet,
-    Map<String, Object> initiallyResolvedHashes
-  ) {
-    Map<String, String> initiallyResolvedAsStrings = new HashMap<>();
-    // This creates a stringified snapshot of the context
-    // so it can be disabled via the config because it may cause performance issues.
-    Stream<Entry<String, Object>> entryStream =
-      (
-        interpreter.getConfig().getExecutionMode().useEagerContextReverting()
-          ? entrySet
-          : interpreter.getContext().getCombinedScope().entrySet()
-      ).stream()
-        .filter(entry -> initiallyResolvedHashes.containsKey(entry.getKey()))
-        .filter(
-          entry -> EagerExpressionResolver.isResolvableObject(entry.getValue(), 4, 400) // TODO make this configurable
-        );
-    entryStream.forEach(
-      entry ->
-        cacheRevertibleObject(
-          interpreter,
-          initiallyResolvedHashes,
-          initiallyResolvedAsStrings,
-          entry
-        )
+    return EagerContextWatcher.executeInChildContext(
+      function,
+      interpreter,
+      eagerChildContextConfig
     );
-    return initiallyResolvedAsStrings;
-  }
-
-  private static Map<String, Object> getInitiallyResolvedHashes(
-    Set<Entry<String, Object>> entrySet,
-    Set<String> metaContextVariables
-  ) {
-    return entrySet
-      .stream()
-      .filter(entry -> !metaContextVariables.contains(entry.getKey()))
-      .filter(
-        entry -> !(entry.getValue() instanceof DeferredValue) && entry.getValue() != null
-      )
-      .collect(
-        Collectors.toMap(Entry::getKey, entry -> getObjectOrHashCode(entry.getValue()))
-      );
-  }
-
-  private static Map<String, Object> getBasicSpeculativeBindings(
-    JinjavaInterpreter interpreter,
-    EagerChildContextConfig eagerChildContextConfig,
-    Set<String> metaContextVariables,
-    Map<String, Object> speculativeBindings
-  ) {
-    speculativeBindings.putAll(
-      interpreter
-        .getContext()
-        .getScope()
-        .entrySet()
-        .stream()
-        .filter(
-          entry ->
-            entry.getValue() instanceof DeferredLazyReferenceSource &&
-            !(((DeferredLazyReferenceSource) entry.getValue()).isReconstructed())
-        )
-        .peek(
-          entry -> ((DeferredLazyReferenceSource) entry.getValue()).setReconstructed(true)
-        )
-        .collect(
-          Collectors.toMap(
-            Entry::getKey,
-            entry -> ((DeferredLazyReferenceSource) entry.getValue()).getOriginalValue()
-          )
-        )
-    );
-    return speculativeBindings
-      .entrySet()
-      .stream()
-      .filter(entry -> !metaContextVariables.contains(entry.getKey()))
-      .filter(entry -> !"loop".equals(entry.getKey()))
-      .map(
-        entry -> {
-          if (
-            eagerChildContextConfig.takeNewValue &&
-            !(entry.getValue() instanceof DeferredValue) &&
-            entry.getValue() != null
-          ) {
-            return entry;
-          }
-          Object contextValue = interpreter.getContext().get(entry.getKey());
-          if (
-            contextValue instanceof DeferredValue &&
-            ((DeferredValue) contextValue).getOriginalValue() != null
-          ) {
-            if (
-              !eagerChildContextConfig.takeNewValue &&
-              !EagerExpressionResolver.isResolvableObject(
-                ((DeferredValue) contextValue).getOriginalValue()
-              )
-            ) {
-              throw new DeferredValueException(
-                String.format("%s: %s", CANNOT_RECONSTRUCT_MESSAGE, entry.getKey())
-              );
-            }
-            return new AbstractMap.SimpleImmutableEntry<>(
-              entry.getKey(),
-              ((DeferredValue) contextValue).getOriginalValue()
-            );
-          }
-          return null;
-        }
-      )
-      .filter(Objects::nonNull)
-      .collect(
-        Collectors.toMap(
-          Entry::getKey,
-          entry ->
-            entry.getValue() instanceof DeferredValue
-              ? ((DeferredValue) entry.getValue()).getOriginalValue()
-              : entry.getValue()
-        )
-      );
-  }
-
-  private static Map<String, Object> getAllSpeculativeBindings(
-    JinjavaInterpreter interpreter,
-    EagerChildContextConfig eagerChildContextConfig,
-    Set<String> metaContextVariables,
-    Map<String, Object> initiallyResolvedHashes,
-    Map<String, String> initiallyResolvedAsStrings,
-    Map<String, Object> speculativeBindings
-  ) {
-    speculativeBindings =
-      speculativeBindings
-        .entrySet()
-        .stream()
-        .filter(
-          entry ->
-            entry.getValue() != null &&
-            !entry.getValue().equals(interpreter.getContext().get(entry.getKey()))
-        )
-        .filter(
-          entry ->
-            !(interpreter.getContext().get(entry.getKey()) instanceof DeferredValue)
-        )
-        .collect(Collectors.toMap(Entry::getKey, Entry::getValue));
-    speculativeBindings.putAll(
-      interpreter
-        .getContext()
-        .entrySet()
-        .stream()
-        .filter(entry -> initiallyResolvedHashes.containsKey(entry.getKey()))
-        .filter(
-          entry ->
-            !initiallyResolvedHashes
-              .get(entry.getKey())
-              .equals(getObjectOrHashCode(entry.getValue()))
-        )
-        .collect(
-          Collectors.toMap(
-            Entry::getKey,
-            entry ->
-              getOriginalValue(
-                interpreter,
-                eagerChildContextConfig,
-                initiallyResolvedHashes,
-                initiallyResolvedAsStrings,
-                entry
-              )
-          )
-        )
-    );
-
-    speculativeBindings =
-      speculativeBindings
-        .entrySet()
-        .stream()
-        .filter(entry -> !metaContextVariables.contains(entry.getKey()))
-        .filter(
-          entry ->
-            !(entry.getValue() instanceof DeferredValue) && entry.getValue() != null
-        ) // these are already set recursively
-        .collect(Collectors.toMap(Entry::getKey, Entry::getValue));
-    return speculativeBindings;
-  }
-
-  private static void cacheRevertibleObject(
-    JinjavaInterpreter interpreter,
-    Map<String, Object> initiallyResolvedHashes,
-    Map<String, String> initiallyResolvedAsStrings,
-    Entry<String, Object> entry
-  ) {
-    RevertibleObject revertibleObject = interpreter
-      .getRevertibleObjects()
-      .get(entry.getKey());
-    Object hashCode = initiallyResolvedHashes.get(entry.getKey());
-    try {
-      if (revertibleObject == null || !hashCode.equals(revertibleObject.getHashCode())) {
-        revertibleObject =
-          new RevertibleObject(
-            hashCode,
-            PyishObjectMapper.getAsPyishStringOrThrow(entry.getValue())
-          );
-        interpreter.getRevertibleObjects().put(entry.getKey(), revertibleObject);
-      }
-      revertibleObject
-        .getPyishString()
-        .ifPresent(
-          pyishString -> initiallyResolvedAsStrings.put(entry.getKey(), pyishString)
-        );
-    } catch (Exception e) {
-      interpreter
-        .getRevertibleObjects()
-        .put(entry.getKey(), new RevertibleObject(hashCode));
-    }
-  }
-
-  private static Object getOriginalValue(
-    JinjavaInterpreter interpreter,
-    EagerChildContextConfig eagerChildContextConfig,
-    Map<String, Object> initiallyResolvedHashes,
-    Map<String, String> initiallyResolvedAsStrings,
-    Entry<String, Object> e
-  ) {
-    if (eagerChildContextConfig.takeNewValue) {
-      if (e.getValue() instanceof DeferredValue) {
-        return ((DeferredValue) e.getValue()).getOriginalValue();
-      }
-      return e.getValue();
-    }
-
-    if (
-      e.getValue() instanceof DeferredValue &&
-      initiallyResolvedHashes
-        .get(e.getKey())
-        .equals(getObjectOrHashCode(((DeferredValue) e.getValue()).getOriginalValue()))
-    ) {
-      return ((DeferredValue) e.getValue()).getOriginalValue();
-    }
-
-    // This is necessary if a state-changing function, such as .update()
-    // or .append() is run against a variable in the context.
-    // It will revert the effects when takeNewValue is false.
-    if (initiallyResolvedAsStrings.containsKey(e.getKey())) {
-      // convert to new list or map
-      try {
-        return interpreter.resolveELExpression(
-          initiallyResolvedAsStrings.get(e.getKey()),
-          interpreter.getLineNumber()
-        );
-      } catch (DeferredValueException ignored) {}
-    }
-
-    // Previous value could not be mapped to a string
-    throw new DeferredValueException(
-      String.format("%s: %s", CANNOT_RECONSTRUCT_MESSAGE, e.getKey())
-    );
-  }
-
-  private static Object getObjectOrHashCode(Object o) {
-    if (o instanceof LazyExpression) {
-      o = ((LazyExpression) o).get();
-    }
-    if (o instanceof PyList && !((PyList) o).toList().contains(o)) {
-      return o.hashCode();
-    }
-    if (o instanceof PyMap && !((PyMap) o).toMap().containsValue(o)) {
-      return o.hashCode() + ((PyMap) o).keySet().hashCode();
-    }
-    return o;
   }
 
   /**
@@ -484,15 +157,15 @@ public class EagerReconstructionUtils {
       .peek(entry -> entry.getValue().setDeferred(true))
       .map(
         entry ->
-          executeInChildContext(
+          EagerContextWatcher.executeInChildContext(
             eagerInterpreter ->
               EagerExpressionResult.fromString(
                 new EagerMacroFunction(entry.getKey(), entry.getValue(), interpreter)
                 .reconstructImage()
               ),
             interpreter,
-            EagerChildContextConfig
-              .newBuilder()
+            EagerContextWatcher
+              .EagerChildContextConfig.newBuilder()
               .withForceDeferredExecutionMode(true)
               .build()
           )
@@ -966,79 +639,5 @@ public class EagerReconstructionUtils {
         false
       );
     return extraStuff;
-  }
-
-  public static class EagerChildContextConfig {
-    private final boolean takeNewValue;
-
-    private final boolean discardSessionBindings;
-    private final boolean partialMacroEvaluation;
-
-    private final boolean checkForContextChanges;
-    private final boolean forceDeferredExecutionMode;
-
-    private EagerChildContextConfig(
-      boolean takeNewValue,
-      boolean discardSessionBindings,
-      boolean partialMacroEvaluation,
-      boolean checkForContextChanges,
-      boolean forceDeferredExecutionMode
-    ) {
-      this.takeNewValue = takeNewValue;
-      this.discardSessionBindings = discardSessionBindings;
-      this.partialMacroEvaluation = partialMacroEvaluation;
-      this.checkForContextChanges = checkForContextChanges;
-      this.forceDeferredExecutionMode = forceDeferredExecutionMode;
-    }
-
-    public static Builder newBuilder() {
-      return new Builder();
-    }
-
-    public static class Builder {
-      private boolean takeNewValue;
-
-      private boolean discardSessionBindings;
-      private boolean partialMacroEvaluation;
-      private boolean checkForContextChanges;
-      private boolean forceDeferredExecutionMode;
-
-      private Builder() {}
-
-      public Builder withTakeNewValue(boolean takeNewValue) {
-        this.takeNewValue = takeNewValue;
-        return this;
-      }
-
-      public Builder withDiscardSessionBindings(boolean discardSessionBindings) {
-        this.discardSessionBindings = discardSessionBindings;
-        return this;
-      }
-
-      public Builder withPartialMacroEvaluation(boolean partialMacroEvaluation) {
-        this.partialMacroEvaluation = partialMacroEvaluation;
-        return this;
-      }
-
-      public Builder withCheckForContextChanges(boolean checkForContextChanges) {
-        this.checkForContextChanges = checkForContextChanges;
-        return this;
-      }
-
-      public Builder withForceDeferredExecutionMode(boolean forceDeferredExecutionMode) {
-        this.forceDeferredExecutionMode = forceDeferredExecutionMode;
-        return this;
-      }
-
-      public EagerChildContextConfig build() {
-        return new EagerChildContextConfig(
-          takeNewValue,
-          discardSessionBindings,
-          partialMacroEvaluation,
-          checkForContextChanges,
-          forceDeferredExecutionMode
-        );
-      }
-    }
   }
 }
