+++ src/test/java/com/hubspot/jinjava/util/EagerReconstructionUtilsTest.java
@@ -25,7 +25,6 @@ import com.hubspot.jinjava.objects.collections.PyMap;
 import com.hubspot.jinjava.tree.TagNode;
 import com.hubspot.jinjava.tree.parse.DefaultTokenScannerSymbols;
 import com.hubspot.jinjava.util.EagerExpressionResolver.EagerExpressionResult;
-import com.hubspot.jinjava.util.EagerReconstructionUtils.EagerChildContextConfig;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashSet;
@@ -67,7 +66,7 @@ public class EagerReconstructionUtilsTest extends BaseInterpretingTest {
   @Test
   public void itExecutesInChildContextAndTakesNewValue() {
     context.put("foo", new PyList(new ArrayList<>()));
-    EagerExecutionResult result = EagerReconstructionUtils.executeInChildContext(
+    EagerExecutionResult result = EagerContextWatcher.executeInChildContext(
       (
         interpreter1 -> {
           ((List<Integer>) interpreter1.getContext().get("foo")).add(1);
@@ -75,8 +74,8 @@ public class EagerReconstructionUtilsTest extends BaseInterpretingTest {
         }
       ),
       interpreter,
-      EagerChildContextConfig
-        .newBuilder()
+      EagerContextWatcher
+        .EagerChildContextConfig.newBuilder()
         .withTakeNewValue(true)
         .withForceDeferredExecutionMode(true)
         .withCheckForContextChanges(true)
@@ -94,7 +93,7 @@ public class EagerReconstructionUtilsTest extends BaseInterpretingTest {
   @Test
   public void itExecutesInChildContextAndDefersNewValue() {
     context.put("foo", new ArrayList<Integer>());
-    EagerExecutionResult result = EagerReconstructionUtils.executeInChildContext(
+    EagerExecutionResult result = EagerContextWatcher.executeInChildContext(
       (
         interpreter1 -> {
           context.put(
@@ -105,8 +104,8 @@ public class EagerReconstructionUtilsTest extends BaseInterpretingTest {
         }
       ),
       interpreter,
-      EagerChildContextConfig
-        .newBuilder()
+      EagerContextWatcher
+        .EagerChildContextConfig.newBuilder()
         .withForceDeferredExecutionMode(true)
         .withCheckForContextChanges(true)
         .build()
@@ -370,26 +369,26 @@ public class EagerReconstructionUtilsTest extends BaseInterpretingTest {
   @Test
   public void itDiscardsSessionBindings() {
     interpreter.getContext().put("foo", "bar");
-    EagerExecutionResult withSessionBindings = EagerReconstructionUtils.executeInChildContext(
+    EagerExecutionResult withSessionBindings = EagerContextWatcher.executeInChildContext(
       eagerInterpreter -> {
         interpreter.getContext().put("foo", "foobar");
         return EagerExpressionResult.fromString("");
       },
       interpreter,
-      EagerChildContextConfig
-        .newBuilder()
+      EagerContextWatcher
+        .EagerChildContextConfig.newBuilder()
         .withDiscardSessionBindings(false)
         .withCheckForContextChanges(true)
         .build()
     );
-    EagerExecutionResult withoutSessionBindings = EagerReconstructionUtils.executeInChildContext(
+    EagerExecutionResult withoutSessionBindings = EagerContextWatcher.executeInChildContext(
       eagerInterpreter -> {
         interpreter.getContext().put("foo", "foobar");
         return EagerExpressionResult.fromString("");
       },
       interpreter,
-      EagerChildContextConfig
-        .newBuilder()
+      EagerContextWatcher
+        .EagerChildContextConfig.newBuilder()
         .withDiscardSessionBindings(true)
         .withCheckForContextChanges(true)
         .build()
